fix quelques bugs

+++ src/main/java/com/cpe/sumulationserver/model/AlertEntity.java
@@ -17,10 +17,10 @@ public class AlertEntity {
     private Timestamp date;
     @Basic
     @Column(name = "longitude")
-    private Float longitude;
+    private Double longitude;
     @Basic
     @Column(name = "latitude")
-    private Float latitude;
+    private Double latitude;
     @Basic
     @Column(name = "intensity")
     private Integer intensity;
@@ -45,19 +45,19 @@ public class AlertEntity {
         this.date = date;
     }
 
-    public Float getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Float longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
 
-    public Float getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Float latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
